+++ src/api/Instance/authApi.ts
@@ -1,5 +1,4 @@
 import axios from "axios";
-import { removeToken } from "../../utils/jwt";
 import { logout } from "../auth";
 
 const API_URL = import.meta.env.VITE_API_URL as string;
@@ -7,14 +6,11 @@ const API_URL = import.meta.env.VITE_API_URL as string;
 const authApi = axios.create({
   baseURL: API_URL,
   timeout: 5000,
+  withCredentials: true,
 });
 
 authApi.interceptors.request.use(
   (config) => {
-    const accessToken: string | null = localStorage.getItem("accessToken");
-    if (accessToken) {
-      config.headers.Authorization = `Bearer ${accessToken}`;
-    }
     return config;
   },
   (error) => {
@@ -28,7 +24,6 @@ authApi.interceptors.response.use(
     const statusCode = error.response?.status;
 
     if (statusCode === 401 || statusCode === 403) {
-      removeToken();
       logout();
       window.location.href = "/login";
     }
