+++ src/api/Instance/baseApi.ts
@@ -5,6 +5,7 @@ const API_URL = import.meta.env.VITE_API_URL as string;
 const baseApi = axios.create({
   baseURL: API_URL,
   timeout: 5000,
+  withCredentials: true,
 });
 
 export default baseApi;
